pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-ingestion"
version = "0.1.0"
description = "Energy invoice extraction pipeline with multi-pass LLM processing"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pymupdf>=1.23",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    "langdetect>=1.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "rich>=13.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "azure-storage-blob>=12.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "httpx",
    "ruff>=0.4",
]

[project.scripts]
invoice-worker = "invoice_ingestion.workers.blob_processor:main"

[tool.hatch.build.targets.wheel]
packages = ["src/invoice_ingestion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring external services (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
